+++ CHANGELOG.md
@@ -8,6 +8,9 @@ have notable changes.
 
 Changes since v2.32
 
+### Fixes
+- Label and header fields have URLs made into links. (#3155)
+
 ## v2.32 "Itämerenkatu" 2023-04-26
 
 **NB: GET /entitlements.csv is deprecated and will be removed in a future release. Please use GET /api/entitlements/export-csv instead.**